// src/app/enums/common-enums.ts
export enum MenuType {
  None = 0,
  Invoice = 1,
  WholesaleInvoice = 2,
  GRN = 3,
  GRNOrder = 4,
  CashHandler = 5,
  Expenses = 6,
  StockTaking = 7,
  ItemMaster = 8,
  ItemTypes = 9,
  Customer = 10,
  SupplierMaster = 11,
  DailySales = 12,
  TimeWiseSales = 13,
  SalesDetail = 14,
  ItemWiseSales = 15,
  MonthlySummary = 16,
  Comments = 17,
  ItemMoving = 18,
}